remove name padding

+++ sw/device/silicon_creator/manuf/base/ft_personalize.c
@@ -158,7 +158,7 @@ static uint8_t all_certs[8192];
 // 1K should be enough for the largest certificate perso LTV object.
 enum { kBufferSize = 1024 };
 static alignas(uint32_t) uint8_t cert_buffer[kBufferSize];
-static alignas(uint32_t) dice_page_t dice_page;
+static alignas(uint64_t) dice_page_t dice_page;
 static size_t uds_offset;
 static size_t cdi_0_offset;
 static size_t cdi_1_offset;
